# cmd/api/docs/swagger.yaml
definitions:
  domain.Address:
    properties:
      addressId:
        type: integer
      city:
        type: string
      country:
        type: string
      district:
        type: string
      houseName:
        type: string
      houseNo:
        type: string
      landmark:
        type: string
      mobile:
        type: string
      name:
        type: string
      pincode:
        type: string
      place:
        type: string
      user:
        $ref: '#/definitions/domain.User'
      userId:
        type: integer
    type: object
...
  domain.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      password:
        type: string
      permission:
        type: boolean
      premium:
        type: boolean
    required:
    - email
    - name
    - password
    type: object
...
  models.Admin:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.ErrorResponse:
    properties:
      err:
        type: string
      hint:
        type: string
      status:
        type: string
    type: object
  models.ListAddress:
    properties:
      addressId:
        type: integer
      city:
        type: string
      country:
        type: string
      district:
        type: string
      houseName:
        type: string
      houseNo:
        type: string
      landmark:
        type: string
      mobile:
        type: string
      name:
        type: string
      pincode:
        type: string
      place:
        type: string
      userId:
        type: integer
    type: object
...
  models.UselistResponse:
    properties:
      filter:
        type: integer
      lastpage:
        type: integer
      list:
        items:
          $ref: '#/definitions/domain.User'
        type: array
      page:
        type: integer
      search:
        type: string
      size:
        type: integer
    type: object
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      permission:
        type: boolean
    type: object
  models.Userlogindata:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /admin/admins:
    get:
      description: Get a list of Admins
      produces:
      - application/json
      responses:
        "200":
          description: List of Admins
          schema:
            items:
              $ref: '#/definitions/models.Admin'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List Admins
      tags:
      - Admin
...
  /admin/blockuser/{id}:
    put:
      description: Block or unblock a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User is blocked/unblocked successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Block or unblock user by ID
      tags:
      - Admin
...
  /admin/user/{id}:
    delete:
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User is deleted successfully
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete user by ID
      tags:
      - Admin
    get:
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Find user by ID
      tags:
      - Admin
  /admin/users:
    post:
      description: |-
        Lists all users from the database.
        jwt temp:  eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6MSwicm9sZSI6ImFkbWluIn0.bw-k7fjfW6nY9DIBZ46ZcJG0IdaYclwHW7P7IPwuNoQ
      parameters:
      - description: 'Page number for pagination (default: 1)'
        in: query
        name: page
        type: integer
      - description: Search word if any
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UselistResponse'
        "400":
          description: BadRequest
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: InternalServerError
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      - Cookie: []
      summary: Get all users
      tags:
      - Admin
  /adminlogin:
    get:
      description: Get the HTML page for admin login
      produces:
      - text/html
      responses:
        "200":
          description: 'got html page: login as admin'
          schema:
            type: string
      summary: Get login page for admin
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Login as an admin
      parameters:
      - description: Admin credentials
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.Userlogindata'
      produces:
      - application/json
      responses:
        "200":
          description: logged in TOKEN_STRING
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Admin Login
      tags:
      - Admin
...
  /user/addaddress:
    post:
      consumes:
      - application/json
      description: Add a new address for a user.
      parameters:
      - description: Address details
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/domain.Address'
      produces:
      - application/json
      responses:
        "200":
          description: Address added
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Add address
      tags:
      - address
...
  /user/deleteaddress:
    delete:
      description: Delete an address by its ID.
      parameters:
      - description: Address ID
        in: query
        name: addressid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Address deleted
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Delete address
      tags:
      - address
...
  /user/getaddress:
    get:
      description: Retrieve an address by its ID.
      parameters:
      - description: Address ID
        in: query
        name: addressid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Address details
          schema:
            $ref: '#/definitions/models.ListAddress'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Get address
      tags:
      - address
...
  /user/listaddresses:
    get:
      description: Retrieve a list of addresses for a user by user ID.
      parameters:
      - description: User ID
        in: query
        name: userid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of addresses
          schema:
            items:
              items:
                $ref: '#/definitions/models.ListAddress'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: List addresses
      tags:
      - address
...
  /user/updateaddress:
    put:
      consumes:
      - application/json
      description: Update an existing address for a user.
      parameters:
      - description: Updated address details
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/domain.Address'
      produces:
      - application/json
      responses:
        "200":
          description: Address updated
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Update address
      tags:
      - address
...
swagger: "2.0"